# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: AI Text-to-Image Generator API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /generate:
    post:
      operationId: generateImage
      tags: [images]
      summary: Generate images from a text prompt
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GenerateImageInput"
      responses:
        "201":
          description: Generation started or completed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GeneratedImage"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /images:
    get:
      operationId: listImages
      tags: [images]
      summary: List generated images (history)
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: search
          in: query
          schema:
            type: ["string", "null"]
        - name: style
          in: query
          schema:
            type: ["string", "null"]
        - name: favorite
          in: query
          schema:
            type: ["boolean", "null"]
        - name: startDate
          in: query
          schema:
            type: ["string", "null"]
        - name: endDate
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: Paginated list of images
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ImageListResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /images/recent:
    get:
      operationId: listRecentImages
      tags: [images]
      summary: Get recent generated images for dashboard
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 8
      responses:
        "200":
          description: Recent images
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GeneratedImage"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /images/{id}:
    get:
      operationId: getImage
      tags: [images]
      summary: Get a specific image
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Image details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GeneratedImage"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    patch:
      operationId: updateImage
      tags: [images]
      summary: Update image metadata (prompt, title)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ImageUpdate"
      responses:
        "200":
          description: Updated image
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GeneratedImage"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    delete:
      operationId: deleteImage
      tags: [images]
      summary: Delete a generated image
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /images/{id}/favorite:
    patch:
      operationId: toggleFavorite
      tags: [images]
      summary: Toggle favorite status of an image
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Updated image
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GeneratedImage"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /images/{id}/variations:
    post:
      operationId: createVariation
      tags: [images]
      summary: Create variations of an existing image
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VariationInput"
      responses:
        "201":
          description: Variation generated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GeneratedImage"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
    GenerateImageInput:
      type: object
      required: [prompt]
      properties:
        prompt:
          type: string
          minLength: 1
        negativePrompt:
          type: ["string", "null"]
        style:
          type: ["string", "null"]
        size:
          type: string
          enum: ["1024x1024", "1536x1024", "1024x1536"]
          default: "1024x1024"
        quality:
          type: string
          enum: ["standard", "high"]
          default: "standard"
        seed:
          type: ["integer", "null"]
    ImageUpdate:
      type: object
      properties:
        prompt:
          type: ["string", "null"]
        title:
          type: ["string", "null"]
    VariationInput:
      type: object
      properties:
        prompt:
          type: ["string", "null"]
        style:
          type: ["string", "null"]
    GeneratedImage:
      type: object
      required: [id, prompt, imageUrl, size, style, isFavorite, generationTimeMs, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: ["string", "null"]
        prompt:
          type: string
        negativePrompt:
          type: ["string", "null"]
        title:
          type: ["string", "null"]
        imageUrl:
          type: string
        size:
          type: string
        style:
          type: ["string", "null"]
        quality:
          type: ["string", "null"]
        seed:
          type: ["integer", "null"]
        isFavorite:
          type: boolean
        generationTimeMs:
          type: integer
        createdAt:
          type: string
    ImageListResponse:
      type: object
      required: [images, total, page, limit]
      properties:
        images:
          type: array
          items:
            $ref: "#/components/schemas/GeneratedImage"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
...