# lib/api-zod/src/generated/types/generateImageInput.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * AI Text-to-Image Generator API
 * OpenAPI spec version: 0.1.0
 */
import type { GenerateImageInputQuality } from './generateImageInputQuality';
import type { GenerateImageInputSize } from './generateImageInputSize';

export interface GenerateImageInput {
  /** @minLength 1 */
  prompt: string;
  /** @nullable */
  negativePrompt?: string | null;
  /** @nullable */
  style?: string | null;
  size?: GenerateImageInputSize;
  quality?: GenerateImageInputQuality;
  /** @nullable */
  seed?: number | null;
}