# lib/api-zod/src/generated/types/imageListResponse.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * AI Text-to-Image Generator API
 * OpenAPI spec version: 0.1.0
 */
import type { GeneratedImage } from './generatedImage';

export interface ImageListResponse {
  images: GeneratedImage[];
  total: number;
  page: number;
  limit: number;
}